move GA to head
