+++ apps/web/src/components/ClientAnalyticsScript/GoogleAnalytics.tsx
@@ -1,0 +1,17 @@
+import Script from 'next/script';
+
+export function GoogleAnalytics() {
+  return (
+    <>
+      <Script async src="https://www.googletagmanager.com/gtag/js?id=G-D1QGEV3B07" />
+      <Script id="google-analytics" strategy="afterInteractive">
+        {`
+          window.dataLayer = window.dataLayer || [];
+          function gtag(){dataLayer.push(arguments);}
+          gtag('js', new Date());
+          gtag('config', 'G-D1QGEV3B07');
+        `}
+      </Script>
+    </>
+  );
+}
